resize: fixed clang-tidy complains

+++ src/capture_filter/resize.c
@@ -258,8 +258,8 @@ static struct video_frame *filter(void *state, struct video_frame *in)
             s->decoder == vc_memcpy ? in : s->dec_frame;
 
         resize_frame(in_frame->tiles[i].data, in_frame->color_spec,
-                     out_frame->tiles[i].data, in_frame->tiles[i].width,
-                     in_frame->tiles[i].height, &s->param);
+                     out_frame->tiles[i].data, (int) in_frame->tiles[i].width,
+                     (int) in_frame->tiles[i].height, &s->param);
     }
 
     VIDEO_FRAME_DISPOSE(in);
